update 첫화면 작성
